refactor(MoveStatus): 클래스명 변경
