--- src/main/java/model/CarStatus.java
@@ -1,25 +1,0 @@
-package model;
-
-public enum CarStatus {
-  FORWARD("-"),
-  STOP("");
-
-  private static final int FORWARD_CONDITION = 4;
-
-  private final String output;
-
-  CarStatus(String output) {
-    this.output = output;
-  }
-
-  public String getOutput() {
-    return output;
-  }
-
-  public static CarStatus decide(int num) {
-    if (num >= FORWARD_CONDITION) {
-      return FORWARD;
-    }
-    return STOP;
-  }
-}
