+++ src/test/java/model/MoveStatusTest.java
@@ -6,12 +6,12 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class CarStatusTest {
+class MoveStatusTest {
   @Test
   @DisplayName("자동차의 전진과 후진을 판단할 수 있다.")
   void decide() {
     assertAll(
-        () -> assertThat(CarStatus.decide(4)).isEqualTo(CarStatus.FORWARD),
-        () -> assertThat(CarStatus.decide(3)).isEqualTo(CarStatus.STOP));
+        () -> assertThat(MoveStatus.decide(4)).isEqualTo(MoveStatus.FORWARD),
+        () -> assertThat(MoveStatus.decide(3)).isEqualTo(MoveStatus.STOP));
   }
 }
